refactor: Rename serial port QEMU driver directory and update includes
