+++ alkos/kernel/arch/x86_64/src/hal/debug_terminal.hpp
@@ -5,7 +5,7 @@
 #include <assert.h>
 #include <autogen/feature_flags.h>
 #include <extensions/defines.hpp>
-#include "drivers/serial_port_qemu/serial_qemu.hpp"
+#include "drivers/serial/qemu.hpp"
 
 namespace arch
 {
